git add "add,edit, delete, view success when create a hotel with hote…

+++ Areas/Owner/Controllers/HotelController.cs
@@ -8,12 +8,11 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.VisualBasic;
+using Microsoft.VisualStudio.Web.CodeGenerators.Mvc.Templates.BlazorIdentity.Pages;
 
 namespace HotelApp.Controllers
 {
     [Area("Owner")]
-    /*[Authorize(Roles = "Owner")]*/ // test with Admin role, because I still not finish Owner Role
-    
     public class HotelController : Controller
     {
         private readonly IHotelRepository HotelRepository;
@@ -37,9 +36,10 @@ namespace HotelApp.Controllers
             List<Hotel> hotels = _dbContext.Hotels.ToList();
             return View(hotels);
         }
-        public IActionResult Hotellist()
+        public async Task <IActionResult> Hotellist()
         {
-            List<Hotel> hotels = _unitOfWork.HotelRepository.GetAll().ToList();
+            var ownerId = _userManager.GetUserId(User);
+            var hotels = await _unitOfWork.HotelRepository.GetHotelByOwnerId(ownerId);
             return View(hotels);
         }
 
@@ -88,6 +88,7 @@ namespace HotelApp.Controllers
             // Model: List room
             // ViewBag: Hotel
         }
+        [HttpGet]
         public IActionResult CreateHotel()
         {
             return View();
@@ -96,6 +97,7 @@ namespace HotelApp.Controllers
         public async Task<IActionResult> CreateHotel(Hotel hotel, IFormFile? imageFile)
         {
             var user = await _userManager.GetUserAsync(User);
+            var ownerId = _userManager.GetUserId(User);
             if (ModelState.IsValid)
             {
                 // Upload image
@@ -119,8 +121,9 @@ namespace HotelApp.Controllers
                         await imageFile.CopyToAsync(fileStream);
                     }
                     hotel.ImageUrl = $"/img/" + fileName;
-
                 }
+
+                hotel.OwnerId = ownerId;
                 if(await _userManager.IsInRoleAsync(user, Constraintt.User) && !await _userManager.IsInRoleAsync(user, Constraintt.Owner))
                 {
                     await _userManager.AddToRoleAsync(user, Constraintt.Owner);
@@ -129,18 +132,25 @@ namespace HotelApp.Controllers
                 _unitOfWork.Save();
                 TempData["success"] = "Hotel created successfully";
                 TempData["ShowMessage"] = true;
-                return RedirectToAction("Index");
+                return RedirectToAction("HotelList");
             }
             return View();
         }
 
-        public IActionResult EditHotel(int? id)
+        [HttpGet]
+        public async Task<IActionResult> EditHotel(int? id)
         {
+            
             if (id == null || id == 0)
             {
-                return NotFound();
+               return NotFound();
             }
-            Hotel? hotel = _unitOfWork.HotelRepository.Get(h => h.HotelId == id);
+            var ownerId = _userManager.GetUserId(User);
+            if (!await _unitOfWork.HotelRepository.IsHotelOwnerAsync(id.Value, ownerId))
+            {
+                return Unauthorized();
+            }
+            var hotel = await _unitOfWork.HotelRepository.GetHotelById(id.Value);
             if (hotel == null)
             {
                 NotFound();
@@ -149,9 +159,18 @@ namespace HotelApp.Controllers
         }
 
         [HttpPost]
-        [Authorize(Roles = "Admin, Owner")]
-        public IActionResult EditHotel(Hotel hotel, IFormFile? imageFile)
+        [Authorize(Roles = "Owner")]
+        public async Task<IActionResult> EditHotel(Hotel hotel, IFormFile? imageFile, int hotelId)
         {
+            if(hotelId != hotel.HotelId)
+            {
+                return NotFound();
+            }
+            var ownerId = _userManager.GetUserId(User);
+            if (!await _unitOfWork.HotelRepository.IsHotelOwnerAsync(hotelId, ownerId))
+            {
+                return Unauthorized();
+            }
             if (ModelState.IsValid)
             {
                 if (imageFile != null && imageFile.Length > 0)
@@ -169,34 +188,46 @@ namespace HotelApp.Controllers
                 _unitOfWork.HotelRepository.Update(hotel);
                 _unitOfWork.Save();
                 TempData["success"] = "Hotel Updated successfully";
-                return RedirectToAction("Index");
+                return RedirectToAction("HotelList");
             }
             return View(hotel);
         }
+
         [HttpGet]
-        public IActionResult DeleteHotel(int? id)
+        public async Task<IActionResult> DeleteHotel(int? id)
         {
             Hotel hotel = new Hotel();
             if (id == 0 || id == null)
             {
                 return NotFound();
             }
-            hotel = _unitOfWork.HotelRepository.Get(h => h.HotelId == id);
+            hotel = await _unitOfWork.HotelRepository.GetHotelById(id.Value);
             if (hotel == null)
             {
-                return NotFound();
+                return NotFound();  
+            }
+            var ownerId = _userManager.GetUserId(User);
+            if (!await _unitOfWork.HotelRepository.IsHotelOwnerAsync(id.Value, ownerId))
+            {
+                return Unauthorized();
             }
             return View(hotel);
         }
+
         [HttpPost]
-        [Authorize(Roles = "Admin, Owner")]
-        public IActionResult DeleteHotel(Hotel hotel)
+        public async Task<IActionResult> DeleteHotel(Hotel hotel)
         {
+            var ownerId = _userManager.GetUserId(User);
+            if(!await _unitOfWork.HotelRepository.IsHotelOwnerAsync(hotel.HotelId, ownerId)) 
+            {
+                return Unauthorized();
+            }
+
             _unitOfWork.HotelRepository.Delete(hotel);
             _unitOfWork.Save();
             TempData["success"] = "Hotel deleted successfully";
             TempData["ShowMessage"] = true;
-            return RedirectToAction("Index");
+            return RedirectToAction("HotelList");
         }
     }
 }
